Fix Bash runner and handle runner errors
Prefixing the first line with "export" before sourcing it with Bash made
it impossible to have multiple variables and values spanning multiple
lines.  Therefore, the `allexport` option is used instead.

Fixing the Bash runner made a new error visible.  While `export foo=a b`
doesn't work but also doesn't trigger an error, `foo=a b` does, and so
it needs to be handled by the runner orchestrator.  Otherwise, the
benchmark would fail in the middle of the processing because of the
`foo=a b` test case.  As a result, this commits also adds error handling
to the utility and main Python scripts